Rename CardComponent
Rationale:
We never name things ...Component. E.g. the NavBar is also not called
NavBarComponent.

We cannot call the CSS class just `card` because that name is also used
by Bootstrap and we would collide with their styling.
